First commit of the power supplies identification and equalization script
Addition of the script that was used at the machine studies of may 2023
in order to identify and equalize the power supplies used in the FOFB system.
This script performs open loop system identification for the power supplies,
calculates the PI gains that are needed in order to equalize their closed loop
responses, translates the continuous model gains obtained to correspondent FPGA gains
and at last performs closed loop identification for the power supplies as a means
to verify if the expected performance is achieved.