Separate the creation of the estimation data from the estimation process
The previous code was building the estimation data objects used in the
estimation processes (the iddata objects) inside the estimation commands.
This structure makes it harder for the user of the script to use/add
validation commands for the estimated models such as 'resid' or 'compare',
it also hinders the use of multiple estimation methods for the same
estimation data set as well as the inclusion of validation data sets in the
model estimation workflow.

The changes proposed in this commit aim at sepatate the estimation data
objects from the estimation methods, wich adresses the aforementioned
issues.