Change the method of estimation of the resistances and inductances
This change introduces a more general method of estimation of the
resistances and inductances of the RL circuit system that is being
used as our current model for the plant.

The previous code was estimating these parameters considering that
the estimated systems attended a more restrictive format for a first
order system with delay (based on the the "tf" command to construct
transfer functions), however, depending on the estimation method
being used or the user needs, the models that are estimated may appear
in different formats (as is the case when the "zpk" command is used
for instance). This new way of calculating these parameters takes
into account only the dc gain of the models and their poles, which
may be obtained from any first order system model format.